worker/requirements.txt
SQLAlchemy==2.0.37
psycopg2-binary==2.9.10
pika==1.3.2
python-dotenv==1.0.1